requirements.txt
livekit-agents==0.12.5
livekit-plugins-openai==0.10.12
livekit-plugins-deepgram==0.6.15
livekit-plugins-silero==0.7.4
python-dotenv==1.0.1
requests==2.31.0
langchain==0.3.13
pinecone-client==2.2.2
openai==1.3.0